Move CanResolveAsync to net.liftweb.util, add a spec.
It's really a general abstraction, and we'll want to reuse it for RestHelper.
